+++ CHANGES.rst
@@ -5,7 +5,9 @@ Changelog
 2.0.0rc6 (unreleased)
 ---------------------
 
-- Nothing changed yet.
+- add booking_code field to IPrenotazione schema
+  update locales
+  [lucabel]
 
 
 2.0.0rc5 (2023-09-05)